Update packaging for storage module
