--- storage/src/integrationTest/groovy/ncei/catalog/config/TestRabbitConfig.groovy
@@ -1,22 +1,0 @@
-package ncei.catalog.config
-
-import org.springframework.amqp.core.Queue
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.Profile
-
-
-@Profile("test")
-@Configuration
-class TestRabbitConfig{
-
-  @Value('${rabbitmq.queue}')
-  String queueName
-
-  @Bean
-  Queue indexQueue() {
-    new Queue(queueName, true, false, false)
-  }
-
-}
