+++ storage/src/integrationTest/groovy/org/cedar/metadata/storage/config/TestRabbitConfig.groovy
@@ -1,0 +1,22 @@
+package org.cedar.metadata.storage.config
+
+import org.springframework.amqp.core.Queue
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Profile
+
+
+@Profile("test")
+@Configuration
+class TestRabbitConfig{
+
+  @Value('${rabbitmq.queue}')
+  String queueName
+
+  @Bean
+  Queue indexQueue() {
+    new Queue(queueName, true, false, false)
+  }
+
+}
