--- storage/src/main/groovy/ncei/catalog/domain/MetadataRecordInterface.groovy
@@ -1,10 +1,0 @@
-package ncei.catalog.domain
-
-//Parent class to Collections, Granules, Schemas
-interface MetadataRecordInterface {
-
-  Map asMap()
-
-  String recordTable()
-
-}
