+++ storage/src/main/groovy/org/cedar/metadata/storage/Application.groovy
@@ -1,0 +1,13 @@
+package org.cedar.metadata.storage
+
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.retry.annotation.EnableRetry
+
+@EnableRetry
+@SpringBootApplication
+class Application {
+  static void main(String[] args) {
+    SpringApplication.run(Application.class, args)
+  }
+}
