+++ storage/src/main/groovy/org/cedar/metadata/storage/domain/MetadataRecord.groovy
@@ -1,0 +1,12 @@
+package org.cedar.metadata.storage.domain
+
+abstract class MetadataRecord implements MetadataRecordInterface {
+
+  Map asMap() {
+    this.class.declaredFields.findAll {
+      !it.synthetic
+    }.collectEntries {
+      [(it.name): this."$it.name"]
+    }
+  }
+}
