+++ storage/src/main/groovy/org/cedar/metadata/storage/domain/MetadataRecordInterface.groovy
@@ -1,0 +1,10 @@
+package org.cedar.metadata.storage.domain
+
+//Parent class to Collections, Granules, Schemas
+interface MetadataRecordInterface {
+
+  Map asMap()
+
+  String recordTable()
+
+}
